% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumeTracer.R
\name{C_hIntegrandDynamicC}
\alias{C_hIntegrandDynamicC}
\alias{C_hIntegrandStaticC}
\title{Integrands for determining the concentration of conservative tracer within
the hyporeheic zone}
\usage{
C_hIntegrandDynamicC(tau, t, m)

C_hIntegrandStaticC(tau, m)
}
\arguments{
\item{tau}{A vector of water ages for which remaining-flow-weighted
concentration is desired.}

\item{t}{Simulation time at which weighted concentration is desired}

\item{m}{An model environment.  See \code{\link{simulateFlumeConcentration}}
for details.}
}
\value{
A vector of values representing past concentration weighted
  (multiplied) by the washout function:

  \code{(W(tau) * C_c(t - tau))}.

  \code{W(tau)} is the washout function (the CCDF of the age distribution of
  aquifer discharge). \code{W(tau)} assumes steady state hydrology and
  determines the fraction of recharge that is still in the aquifer at water
  age \code{tau}.

  \code{C_hIntegrandStaticC} assumes \code{C_c} is a constant at the value of
  \code{C_c} prior to the release.  For any model iteration,
  \code{C_hIntegrandStaticC} should be integrated from the current model time
  to \code{tau_n}, but only when current model time is less than
  \code{tau_n}.

  \code{C_hIntegrandDyanmic} looks backwards toward the time of release and
  estimates \code{C_c(t - tau)} using linear interpolation between known
  values of \code{C_c} at each time step.  For any model iteration,
  \code{C_hIntegrandDyanmic} should be integrated from \code{tau_0} to
  \code{min(current model time, tau_n)}.
}
\description{
If t represents simulation time, and t=0 is the time of the tracer release,
and tau is a water age of interest, these functions can be numerically
integrated to determine sum of past concentration at time t - tau, weighted
(multiplied) by the washout function (W(tau) -- the fraction of recharge
remaining in the hyporheic zone at water age tau).
}
\details{
When the sum of the integrals of these functions are divided by the integral
of the washout function (W(tau)) from tau_0 to tau_n, the resulting value is
the mean concentration the hyporehic water.  See
\code{\link{optimizationError}} for details.
}
