% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumeTracer.R
\name{createFlume}
\alias{createFlume}
\title{Create a flume model environment}
\usage{
createFlume(...)
}
\arguments{
\item{...}{Named parameter that will be used to create variable in an
environment returned to the user.}
}
\value{
An environment containing variables with names and values determined
  by the parameters passed to \code{...}.
}
\description{
Creates a flume model environment for use with
\code{\link{simulateFlumeConcentration}}
}
\details{
Pass named parameters to \code{...}.  Each parameter will be created as a
variable of the same name in an environment.  Pass one parameter for each
variable required by \code{\link{simulateFlumeConcentration}}. See
\code{\link{simulateFlumeConcentration}} documentation for description of
required parameters and see example under
\code{\link{simulateFlumeConcentration}} for example of using
\code{createFlume}
}
