% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumeTracer.R
\name{optimizationError}
\alias{optimizationError}
\title{Determine error associated with an estimate of channel concentration.}
\usage{
optimizationError(C_c, m, breaks, staticIntegral)
}
\arguments{
\item{C_c}{An estimate of channel concentration in the next model time step.
Values are generated by nlm() to minimim the return value of
\code{optimizationError}.}

\item{m}{The model environment, which must contain variables \code{tau_0,
tau_n, alpha, idx, times,} and \code{C_c}.  \code{idx} is used as a vector
index, and points at values associated with the current model time.
\code{idx} = (current model iteration + 1), because \code{t=0} is
associated with \code{idx=1}.  \code{times} is a vector of simulation times
corresponding to time steps and \code{C_c} is a vector of channel
concentrations.}

\item{breaks}{See \code{\link{pIntegrate}}}

\item{staticIntegral}{The integral of \code{\link{C_hIntegrandStaticC}}
(pre-release concentration weighted by W(tau), the washout function) from
the current simulation time to tau_n, or zero if current simulation time is
greater than tau_n.}
}
\description{
Objective function for an implicit solution to determine the channel
concentration.  The measure of error returned is the difference between: 1)
the mean concentration of the channel and hyporheic zone, weighted by water
storage in each, at time t; and 2) the expected final concentration of water
in the flume.
}
\details{
\code{optimazationError} sums \code{staticIntegral} with integration of
\code{\link{C_hIntegrandDynamicC}} from tau_0 to min(currentModelTime,
tau_n).  This way, the numerical integration doesn't span the discontinuity
in concentration associated with the slug addtion.  The two summed values are
then divided by the integral of W(tau) from tau_0 to tau_n, the washout
function to convert the sum to the mean conservative tracer concentraion in
the hyporehic zone.
}
