% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumeTracer.R
\name{pIntegrate}
\alias{pIntegrate}
\title{Numerical integration with manual subdivisons.}
\usage{
pIntegrate(funs, breaks, ...)
}
\arguments{
\item{funs}{A single function or a list of functions that will be used for
each interval between \code{breaks}.  A list of multiple functions will be
recycles for the intervals, with a warning if number of intervals between
breaks is not an even multiple of length of the list of functions.  Summing
the integral of multiple functions can be useful if, for instance, there is
a discontinuity in the integrand, with one function describing the integrand
to the left of the discontinuity, and the other describing the integrand to
the right of the discontinuity.  In such a case, one of the breaks should
be the x value of the discontinuity!}

\item{breaks}{A vector of values of representing the divisions among
consecutive intervals to be integrated. The length of \code{breaks} is
therefore the number of desired intervals + 1.  For instance, if
\code{breaks} = c(1,2,5,10), \code{pIntegrate} will integration \code{funs}
from 1 to 2, from 2 to 5, and from 5 to 10, and then sum the results. If
there are only two values in the "breaks" list, this function is identical
to integrate.}

\item{...}{pIntegrate calls \code{\link{integrate}} for each subdivision.
Values in '...' are passed to \code{\link{integrate}} unchanged, for each
subdivision.}
}
\description{
A function that integrates by summing the numerical integration across
multiple consecutive ranges predetermined by the user.
}
\details{
If the shape of an integrand is known ahead of time, choosing specific breaks
can ease or speed the integration.
}
