% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumeTracer.R
\name{postProcess}
\alias{postProcess}
\title{Post processing variables}
\usage{
postProcess(m)
}
\arguments{
\item{m}{Model environment that has been executed with
\code{\link{simulateFlumeConcentration}}}
}
\value{
Returns TRUE silently.  As a side effect, creates a \code{C_up} time series
in the \code{m} environment representing the concentration of upwelling water.
}
\description{
Add addtional hydrology descriptors by post-processing the model result.
}
